requirement.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-simpleui==3.9
djangorestframework==3.11.0
Markdown==3.2.1
pytz==2019.3
sqlparse==0.3.0
